Added the ability to handle extended type like %lli for long long integer. Some tests are needed to find out hidden and vicious bugs.

+++ lib/my/my_printf/src/utils/print.c
@@ -9,7 +9,7 @@
 #include "../my_printf.h"
 #include "flags.h"
 
-void put_char_base(long number, int upper_case)
+void put_char_base(long long number, int upper_case)
 {
     if (upper_case) {
         if (number >= 10)
